requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
tensorflow==2.12.0
opencv-python==4.7.0.72
imutils==0.5.4
mediapipe==0.10.1
python-socketio==5.8.0
eventlet==0.33.3
websocket-client==1.6.1
websockets==11.0.3
aiohttp==3.8.4
asyncio==3.4.3